AG-12600 Add line labels dialog icons and move icons to new file
